\name{ouzelr-package}
\alias{ouzelr-package}
\alias{ouzelr}
\docType{package}
\title{
\packageTitle{ouzelr}
}
\description{
\packageDescription{ouzelr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ouzelr}
\packageIndices{ouzelr}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{ouzelr}

Maintainer: \packageMaintainer{ouzelr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
